:sparkles: user list SSR 적용
