+++ src/components/organisms/NavBar/UserList/HydratedUserList.tsx
@@ -1,0 +1,16 @@
+import { Hydrate, dehydrate } from '@tanstack/react-query'
+import getQueryClient from '@/lib/queryClient'
+import { getAllUsers } from '@/services/user'
+import UserList from './UserList'
+
+export default async function HydratedUserList() {
+  const queryClient = getQueryClient()
+  await queryClient.prefetchQuery(['getAllUsers'], getAllUsers)
+  const dehydratedState = dehydrate(queryClient)
+
+  return (
+    <Hydrate state={dehydratedState}>
+      <UserList />
+    </Hydrate>
+  )
+}
