+++ src/lib/queryClient.ts
@@ -1,0 +1,5 @@
+import { cache } from 'react'
+import { QueryClient } from '@tanstack/react-query'
+
+const getQueryClient = cache(() => new QueryClient())
+export default getQueryClient
